requirements.txt
azure-cosmos==4.2.0
azure-functions==1.7.2
azure-servicebus==7.3.3
pytest==6.2.5
dataclasses-jsonschema==2.14.1